refactor: remove links from the menu
