+++ src/pages-data.js
@@ -5,6 +5,7 @@ export default [
   {
     title: 'Sacola',
     path: '/checkout/cart',
+    disabled: true,
     page: Loadable({
       loader: () => import(/* webpackChunkName: 'checkout-cart' */ './pages/checkout-cart'),
       loading: SkeletonPage
@@ -13,6 +14,7 @@ export default [
   {
     title: 'Pagamento',
     path: '/checkout/cart/payment',
+    disabled: true,
     page: Loadable({
       loader: () => import(/* webpackChunkName: 'checkout-payment' */ './pages/checkout-payment'),
       loading: SkeletonPage
